+++ packages/swapkit/helpers/src/types/wallet.ts
@@ -30,6 +30,7 @@ export enum WalletOption {
   BRAVE = "BRAVE",
   COINBASE_MOBILE = "COINBASE_MOBILE",
   COINBASE_WEB = "COINBASE_WEB",
+  CTRL = "CTRL",
   EIP6963 = "EIP6963",
   EXODUS = "EXODUS",
   KEEPKEY = "KEEPKEY",
@@ -49,7 +50,6 @@ export enum WalletOption {
   TALISMAN = "TALISMAN",
   TRUSTWALLET_WEB = "TRUSTWALLET_WEB",
   WALLETCONNECT = "WALLETCONNECT",
-  XDEFI = "XDEFI",
 }
 
 export enum LedgerErrorCode {