+++ packages/plugins/kado/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @swapkit/plugin-kado
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies [[`00488c6`](https://github.com/thorswap/SwapKit/commit/00488c6235e4d1c322cd5b4b26dc8143f90e73d3)]:
+  - @swapkit/helpers@2.0.0
+  - @swapkit/api@2.0.2
+
+## 0.1.2
+
+### Patch Changes
+
+- [#1112](https://github.com/thorswap/SwapKit/pull/1112) [`1859a1d`](https://github.com/thorswap/SwapKit/commit/1859a1d7664ee07d1e23ed406b4ee59669122cec) Thanks [@towanTG](https://github.com/towanTG)! - Fixes kado widget mapping
+
+## 0.1.1
+
+### Patch Changes
+
+- [#1110](https://github.com/thorswap/SwapKit/pull/1110) [`8e64b40`](https://github.com/thorswap/SwapKit/commit/8e64b40af1a97f9d3182ba8b9acb2ff1df73d0a8) Thanks [@towanTG](https://github.com/towanTG)! - Adds missing swap method to plugin
+
 ## 0.1.0
 
 ### Minor Changes