+++ packages/toolboxes/radix/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @swapkit/toolbox-radix
 
+## 1.2.24
+
+### Patch Changes
+
+- Updated dependencies [[`00488c6`](https://github.com/thorswap/SwapKit/commit/00488c6235e4d1c322cd5b4b26dc8143f90e73d3)]:
+  - @swapkit/helpers@2.0.0
+
 ## 1.2.23
 
 ### Patch Changes