+++ packages/wallets/evm-extensions/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @swapkit/wallet-evm-extensions
 
+## 1.1.28
+
+### Patch Changes
+
+- Updated dependencies [[`00488c6`](https://github.com/thorswap/SwapKit/commit/00488c6235e4d1c322cd5b4b26dc8143f90e73d3)]:
+  - @swapkit/helpers@2.0.0
+  - @swapkit/toolbox-evm@1.7.8
+
 ## 1.1.27
 
 ### Patch Changes