+++ packages/wallets/polkadotjs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @swapkit/wallet-polkadotjs
 
+## 1.0.50
+
+### Patch Changes
+
+- Updated dependencies [[`00488c6`](https://github.com/thorswap/SwapKit/commit/00488c6235e4d1c322cd5b4b26dc8143f90e73d3)]:
+  - @swapkit/helpers@2.0.0
+  - @swapkit/toolbox-substrate@1.3.8
+
 ## 1.0.49
 
 ### Patch Changes